validation on backend for extra security

+++ backend/controllers/customerController.js
@@ -60,6 +60,25 @@ const customerController = {
 
   deleteCustomer: async (req, res) => {
     try {
+      const customerToDelete = await Customer.findById(req.params.id).select('reservations').populate;
+      
+      let canBeDeleted = true
+
+      const { reservations } = await Customer.findById(req.params.id).select('reservations').populate('reservations.tripId', 'reservations').lean()
+      const result = reservations.map(r => {
+        const {reservationId, tripId} = r
+        const filtered = tripId.reservations.find(r => r._id.equals(reservationId) && r.state != 'Cancelled')
+        if(filtered !== undefined) {
+          canBeDeleted = true
+        }
+        return filtered
+      })
+
+      if(!canBeDeleted) {
+        res.status(500).json({ message: "Cannot delete customer with reservations" });
+        return
+      }
+
       const deletedCustomer = await Customer.findByIdAndDelete(req.params.id);
       if (!deletedCustomer) {
         res.status(404).json({ message: "Customer not found" });
